fixed things from pr and added basic db struct for session_request
